# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insta Clone API",
        "description": "Documentação da API do clone de Instagram",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "maria.silva"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Maria Silva"
                    },
                    "username": {
                        "type": "string",
                        "example": "maria.silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@example.com"
                    },
                    "avatar_url": {
                        "type": "string",
                        "example": "http://localhost/storage/avatars/foo.jpg"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Fotógrafa e viajante.",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "joined_at": {
                        "type": "string",
                        "example": "10/08/2026"
                    },
                    "posts_count": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "followers_count": {
                        "type": "integer",
                        "example": 340,
                        "nullable": true
                    },
                    "following_count": {
                        "type": "integer",
                        "example": 180,
                        "nullable": true
                    },
                    "is_following": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserBasic": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "profile_photo": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "caption": {
                        "type": "string",
                        "example": "Pôr do sol de hoje 🌅",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "media_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 24
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "liked_by_me": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserBasic"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "body": {
                        "type": "string",
                        "example": "Que foto linda!"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Message": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}